mudando calculo da regulacao por planejamento

+++ src/main/groovy/br/com/m2msolutions/copiloto/modelo/viagem/momento/MomentoViagem.groovy
@@ -10,7 +10,7 @@ import groovy.transform.PackageScope
 class MomentoViagem {
 
     Date momentoDaTransmissao
-    Float percentualDeConclusao
+    BigDecimal percentualDeConclusao
     Veiculo veiculo
 
     private String linhaId
